pyproject.toml
[project]
name = "AxTBot-Public"
version = "2.1.0+11455"
description = "AxTBot 社区机器人是由 AxT社区 开发并维护的公共、公益性QQ官方机器人，基于QQ开放平台 ，使用 Python 和 FastAPI 构建"
authors = [
    {name = "Shanshui2024", email = "shanshui@shanshui.site"},
]
dependencies = [
    "aerich==0.9.1","aiosqlite==0.21.0","annotated-types==0.7.0","anyio==4.10.0",
    "asyncclick==8.1.8.0","bcrypt==4.3.0","cffi==1.17.1","click==8.2.1","colorama==0.4.6",
    "coloredlogs==15.0.1","dictdiffer==0.9.0","fastapi==0.116.1","h11==0.16.0","humanfriendly==10.0",
    "idna==3.10","iso8601==2.1.0","itsdangerous==2.2.0","prompt-toolkit==3.0.51","pycparser==2.22",
    "pydantic==2.11.7","pydantic-core==2.33.2","PyNaCl==1.5.0","pypika-tortoise==0.6.1",
    "pyreadline3==3.5.4","pytz==2025.2","PyYAML==6.0.2","questionary==2.1.0","sniffio==1.3.1",
    "starlette==0.47.2","tortoise-orm==0.25.1","typing-inspection==0.4.1","typing-extensions==4.14.1",
    "uvicorn==0.35.0","wcwidth==0.2.13","requests", "tomli", "aiohttp"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "AGPL-3.0 license"}

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
packages = ["app", "Utils"]
package-dir = {"Utils" = "src/Utils", "app" = "src/app"}
py-modules = ["main"]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov",
    "pytest-mock",
    "requests-mock",
    "httpx"
]
standard = [
    "psutil"
]





[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--verbose --cov=src --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = [
    ".",  # 添加当前目录
]